fix : Renamed sections files to kebab-case
