+++ src/components/sections/faq-section.tsx
