+++ src/components/sections/hero-section.tsx
