+++ src/components/sections/pricing-section.tsx
